+++ src/main/java/de/maxhenkel/openhud/config/ClientConfig.java
@@ -11,6 +11,7 @@ public class ClientConfig {
     public ModConfigSpec.DoubleValue hudWidth;
     public ModConfigSpec.DoubleValue hudScale;
     public ModConfigSpec.BooleanValue renderCardinalDirections;
+    public ModConfigSpec.BooleanValue renderCenterIndicator;
     public ModConfigSpec.IntValue lineColor;
     public ModConfigSpec.EnumValue<SortOrder> waypointSortOrder;
 
@@ -18,6 +19,7 @@ public class ClientConfig {
         hideHud = builder.comment("If the HUD should be hidden").define("hide_hud", false);
         renderWaypointNames = builder.comment("If waypoint names should be rendered on the HUD").define("render_waypoint_names", true);
         renderCardinalDirections = builder.comment("If cardinal directions should be rendered on the HUD").define("render_cardinal_directions", true);
+        renderCenterIndicator = builder.comment("If an indicator should be rendered at the center of the HUD").define("render_center_indicator", false);
         hudFov = builder.comment("The FOV of the HUD").defineInRange("hud_fov", 180D, 12.5D, 360D);
         hudWidth = builder.comment("The width of the HUD").defineInRange("hud_width", 0.5D, 0.1D, 1D);
         hudScale = builder.comment("The scale of the HUD").defineInRange("hud_scale", 1D, 0.1D, 10D);